requirements.txt
streamlit
openai
python-dotenv
python-docx
python-pptx
pymupdf
PyPDF2
pandas
numpy
sentence-transformers
